Yu-Gi-Oh! Forbidden Memories | Darknite/Nitemare
I've been working on this Midi all night last night such as lowering velocities, increasing volumes, trying to make it DK64 friendly as possible. This afternoon I realized some panning events were also missing at loop so that is now fixed and I have turned the volume up on the track a little bit. If this is still uncomfortable for the ears, I definitely need more information on what needs to be fixed and what not because I've worked on this for 3 hours last night and now just woke up and worked on it for another hour.